move ui elements to new directory
