+++ sampleapp/src/main/java/io/matthewnelson/sampleapp/App.kt
@@ -70,6 +70,7 @@ import android.app.Application
 import android.content.Context
 import android.os.Process
 import androidx.core.app.NotificationCompat
+import io.matthewnelson.sampleapp.ui.MainActivity
 import io.matthewnelson.topl_core_base.TorConfigFiles
 import io.matthewnelson.topl_service.TorServiceController
 import io.matthewnelson.topl_service.notification.ServiceNotification
